fix(broker): prove decode rejects at WaxTests and parse VersionRelation from the enum
Move memory_id "12" and relation "nope" fail-closed tests into
BrokerCommandDecodeTests and delete the MCPServer-gated duplicate file.
Parse VersionRelation from wireName; omitted fact_assert relation defaults to .sets.
