+++ Sources/Wax/Broker/BrokerDomainDecode.swift
@@ -54,13 +54,9 @@ extension BrokerCommand {
     }
 
     package static func parseVersionRelation(_ raw: String) throws -> VersionRelation {
-        switch raw.trimmingCharacters(in: .whitespacesAndNewlines).lowercased() {
-        case "sets": return .sets
-        case "updates": return .updates
-        case "extends": return .extends
-        case "retracts": return .retracts
-        default:
+        guard let relation = VersionRelation(wireName: raw) else {
             throw BrokerValidationError.invalid("relation must be one of: sets, updates, extends, retracts")
         }
+        return relation
     }
 }
